requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
dataframe_image==0.2.7
defusedxml==0.7.1
fastjsonschema==2.21.2
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
jupyterlab_pygments==0.3.0
lxml==6.0.2
MarkupSafe==3.0.3
mistune==3.2.0
more-itertools==10.8.0
msal==1.34.0
multidict==6.7.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.0.0
platformdirs==4.5.1
playwright==1.57.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.1
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
soupsieve==2.8.1
tinycss2==1.4.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
webencodings==0.5.1
yarl==1.22.0
